XML ERROR.
Fix the details view to fix data that is in string form when it should be a number (you could also fix this in the database).

Adding a product/editing a product should save the data with the correct type. If it's a number, it would save it as a string.