XML MISSING (x2). FOCUS IN XML.
I want to be able to generate quotes that are higher than the requirements (useful if you get a better list price by slightly increasing the quantity ordered), and for the page to tell me, for example, if you can get a better price if you increase the quantity of the product by X, the current price range vs. the upper price range.

Location: /requisitions/FTWXeHhugKcWyacJjhmt, Request quotations (button)